Rename files within src/game/
