+++ src/game/__init__.py
@@ -2,9 +2,9 @@
 This package contains implementation of rules of the game Mathematico, that is
 the game itself, evaluation and also the arena for playing.
 """
-from ._board import Board
-from ._eval import evaluate, Points
-from ._mathematico import Player, Mathematico, Arena
+from .board import Board
+from .eval import evaluate, Points
+from .mathematico import Player, Mathematico, Arena
 
 
 __all__ = [
